Audio-Bot/requirements.txt
pyaudio
distutils
google-generativeai
SpeechRecognition
pyttsx3
setuptools
keyboard